Setup Analysis dashboard page
