--- src/newViews/AnalysisModule/AnalysisEditModal/PillarAnalysisRow/styles.scss
@@ -1,20 +1,0 @@
-.pillar-analysis-row {
-    display: flex;
-    align-items: flex-start;
-    margin: 0 calc(-2 * var(--dui-spacing-small));
-    padding: var(--dui-spacing-medium) 0;
-    width: calc(100% + 2 * var(--dui-spacing-small));
-
-    .input {
-        margin: 0 var(--dui-spacing-small);
-    }
-
-    .remove-button {
-        margin: 0 var(--dui-spacing-small);
-        margin-top: calc(var(--dui-font-size-small) + 2 * var(--dui-spacing-small));
-    }
-}
-
-.options-popup {
-    width: 480px!important;
-}
